Extract Method Refactor applied and Move Method
Since the state of the movie has to do with its price code,
we started by creating the Price class with Extract
Class. Then we used Move Method to move the getters
of the amount and points from Movie into the Price class.
The inner enumerated Code class was also moved from
the Movie class to the Price class.